Be way less clever
Now it works! What was going on is that `onAppear()` only gets called
once on macOS, because it only appears once. This means that it wasn't
trying to laod the other views.

But I decided that I was being silly. The number of creatures is small.
Now it transfers all of them at the start and caches it in memory. This
allows the `CreatureDetail` view to be just handed the `Creature`
object, rather than have to go fetch it.