Bug fix: Don't waste CPU cycles needlessly updating an Obserable thing
The `values` array is Observable. Any change to it cascades out, and if
there's UI elements on the screen that reference it, that's a waste of
CPU and battery updating a thing that might actually have been updated.

Before updating the array, make sure the value is actually different.
(Which would, in fact, warrant an update!)