Reorganize the `Source` directory
This change spilts the code up into three parts. One is the `Common`
directory, which contains the files that are shared by the CUI and CLI,
one called `Creature Console` that is the GUI app, and `CreatureCLI`,
which is the CLI app.

I broke it out this way to see if I can get the CLI working on Linux.