Start the first of many iOS 18 / macOS 15 changes
I didn't know that there's a `task` I can use on a view to accomplish
work, rather than having to use `onAppear`. `task` is nice in that Swift
makes sure it only runs once, and will stop it when the view is
destroyed.

I also started turning on the Swift 6 errors (new this week with WWDC).
There are a lot of changes that will need to happen to be able to
compile with Swift 6.